# Don't touch UOps.DEFINE_GLOBAL