@geohot 2024-03-02
I want that fixed in a different way, I explicitly want define global removed. The UOps should form a graph where all toposorts are valid.

Do you have a minimal example of a trigger for this failure?