@geohot 2024-03-02
Ahh, I see, it's removed because it's multiplied by zero. Add this as a test in CI (as generic as you can) and I'll merge this PR and deal with it later.