@uuuvn 2024-03-02
> Really we should be tracking buffers and removing them.

I thought about doing that but I don't understand backend of tinygrad good enought to do it just yet.

> Alternatively, we can detect this multiply by zero in lazy.py and not add the node there.

Multiplying by zero is just fist thing i come up with to reproduce an issue outside of llama, this isn't what happening in llama (see discord). This will always be the problem when UOps.DEFINE_GLOBAL is removed not from the end.